refactor: rename artist model directory
