+++ src/artist/models/artist.model.ts
